Move APILookup to functions dir
